fastapi/fastapi-promethous-grafana/src/pyproject.toml
[tool.poetry]
name = "fastapi-promethous-grafana"
version = "0.1.0"
description = ""
authors = ["Seunggwan Song <yanhis97531@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ruff = "^0.1.4"
fastapi = "^0.104.1"
gunicorn = "^21.2.0"
uvicorn = "^0.24.0.post1"
prometheus-fastapi-instrumentator = "^6.1.0"
sqlalchemy = "^1.4.50"
dependency-injector = "^4.41.0"
pytz = "^2023.3.post1"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
databases = {extras = ["asyncpg"], version = "^0.8.0"}
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
